--- Assets/Scripts/Home/SteamManager.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: b603aac145cdb6f48b60777a35ccb116
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
